fix: update import paths after app/shared module split
develop (#12814) moved pipes and websocket service out of app/shared
into app/foundation. Update all affected imports in the global-search
iris feature files:

  app/shared/pipes/artemis-translate.pipe  → app/foundation/pipes/...
  app/shared/pipes/html-for-markdown.pipe  → app/foundation/pipes/...
  app/shared/service/websocket.service     → app/foundation/service/...

Also remove the now-unused SearchEntityType import from the navigation
view component (activeFilters is typed as string[] at this commit).
